feature: implemented /src/hooks directory into README.md for better documentation, removed unused files to keep structure more clearly
