+++ client/src/components/admin/CategoriesManagement/Type/TypeEdit/TypeEdit.jsx
@@ -2,7 +2,7 @@ import ErrorText from '@components/admin/ErrorText/ErrorText';
 import SuccessMessage from '@components/admin/SuccessMessage/SuccessMessage';
 import Spinner from '@UI/Spinner/Spinner';
 import { yupResolver } from '@hookform/resolvers/yup';
-import useIdFromUrl from '@hooks/useId';
+import useIdFromUrl from '@hooks/useIdFromUrl';
 import { errorTypeById, loadTypeById, loadTypeEdit, responseTypeById, responseTypeEdit } from '@redux/admin/selectors';
 import { editType, getTypeCategoryById } from '@redux/admin/type/service';
 import setFormValues from '@utils/setFormValue';
