+++ client/src/components/admin/CategoriesManagement/Type/TypeList/TypeCategoryItem/TypeCategoryItem.jsx
@@ -1,5 +1,5 @@
 import DeletePopUp from '@components/admin/DeletePopUp/DeletePopUp';
-import useScreenSizes from '@hooks/useScreenSizes';
+import useScreenSizes from '@hooks/useScreenSizes/useScreenSizes';
 import { deleteType } from '@redux/admin/type/service';
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
