--- client/src/hooks/useBreadcrumb.jsx
@@ -1,13 +1,0 @@
-import { useLocation } from 'react-router-dom';
-
-export const useBreadcrumbs = () => {
-  const location = useLocation();
-  const pathnames = location.pathname.split('/').filter((x) => x);
-
-  const breadcrumbs = pathnames.map((value, index) => {
-    const to = `/${pathnames.slice(0, index + 1).join('/')}`;
-    return { name: value, path: to };
-  });
-
-  return breadcrumbs;
-};
