+++ client/src/pages/CatalogPage/FilterCategory/FilterByCategory.jsx
@@ -4,7 +4,7 @@ import EmptyText from '@UI/EmptyText/EmptyText';
 import Spinner from '@UI/Spinner/Spinner';
 import { loadCategories, selectCategories } from '@redux/selectors';
 
-import useScreenSizes from '@hooks/useScreenSizes';
+import useScreenSizes from '@hooks/useScreenSizes/useScreenSizes';
 import CategoryCard from './CategoryCard/CategoryCard';
 import SliderBoxMain from '@components/UI/SliderBoxMain/SliderBoxMain';
