+++ client/src/router/layouts/MainLayout/MainLayout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import Header from '@components/Header/Header';
 import Footer from '@components/Footer/Footer';
-import useScrollToTop from '@hooks/useScrollToTop';
+import useScrollToTop from './hooks/useScrollToTop';
 import cl from './index.module.scss';
 import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
